+++ pkg/tlscert/webhookconfig.go
@@ -1,4 +1,4 @@
-package admission
+package tlscert
 
 import (
 	"bytes"
